test: add tests for missing ADUser case
